# perf: Access grid only once