In our current implementation we access the grid first to check if the cell is empty and then a second time to iterate over the contents. Inlining the check and using the walrus operator allows to access the content only once. 

Using the ABM framework comparison i get the following speedups on my laptop for the Schelling model.

| Model | main| this PR|
|--------|--------|--------|
| Schelling (small) | 136ms | 97ms |
| Schelling (large) | 2120ms| 1558ms |