@Corvince 2023-08-03
On the Schelling performance I also opened https://github.com/JuliaDynamics/ABM_Framework_Comparisons/issues/57 which could give us a 3-5x advantage